+++ arduino/chomp/hold_down.h
@@ -3,4 +3,4 @@ void manualHoldDown(bool enable);
 bool autoHoldDown();
 void endHoldDownSample();
 void restoreHoldDownParameters();
-void setHoldDownParameters(int16_t sample_period);
+void setHoldDownParameters(int32_t sample_period);
